refactor(ai-core): split the composer suites and share their harness
The composer/suggestions work left three test files repeating the same
~110-line preamble and one 837-line suite covering four unrelated seams.

Tests:
- add __tests__/support/ (jsdom globals, local-agent runtime double,
  session store factory, mount/DOM helpers). The unstable_mockModule call
  stays per-file, since the ESM registry is per-file and the mocked
  specifier resolves relative to the registering file; only the mutable
  runtime moves, behind setMockRuntime(), which returns the runtime so
  tests assert on a local instead of module-level mutable state
- split ChatComposer.test.tsx into .keymap, .sendStop, .asChild and
  .session suites, grouped by the seam under test
- turn the four-case guard loop into it.each so each case reports by name

Fix three assertions that used toHaveBeenCalledWith(undefined), which
requires one argument equal to undefined, where composer.send() passes
none. The production behavior was already correct.

Source:
- add createDualModeChatContext(), replacing the five-part scaffold that
  was copy-pasted across both mode-normalized contexts, and centralizing
  the reason the boundary dispatches instead of the reader hook
- collapse the two suggestions mode hooks behind a SuggestionsSource
  holding only the values the modes disagree on, so toggle/fill/send are
  derived once

Public API and export names are unchanged.

Also rename ChatActiveStatus.test.ts to .tsx: it imports a .tsx module,
which trips TS6142 in an editor's inferred project because test files sit
outside the package tsconfig's include.
