+++ packages/ai-core/__tests__/ChatActiveStatus.test.tsx
