+++ packages/ai-core/__tests__/support/index.tsx
@@ -1,0 +1,11 @@
+/**
+ * Shared harness for the composer and suggestions suites.
+ *
+ * Importing this barrel installs the jsdom globals as a side effect, so a test
+ * file needs one import for the whole environment.
+ */
+import './env';
+
+export * from './localAgentRuntime';
+export * from './render';
+export * from './sessionStore';
